split bmds session
